관리자 레이어링 리스트, 삭제 추가 // 관리자 footer 에 user 추가

+++ src/main/webapp/WEB-INF/views/admin/user/history/extagram.jsp
@@ -12,7 +12,7 @@
 					<br>
 
 
-<h1 style="text-align: center;"> ${user.nick } 님 Extagram 작성 게시글</h1>
+<h1 style="text-align: center;"> ${user.nick } 님의 Extagram 게시글</h1>
 <hr>
 <div style="text-align: center;">
 <a href="/admin/user/history/extagram?userNo=${user.userNo }"><button>extagram</button></a>
@@ -39,7 +39,7 @@
 		<td style="text-align: center;">${exta.EX_CONTENT }</td>
 		<td style="text-align: center;">${exta.EX_DATE  }</td>
 		<td style="text-align: center;"><a href="/extagram/view?exNo=${exta.EX_NO }"><button>상세보기</button></a></td>
-		<td style="text-align: center;"><a href="/admin/user/history/extagram/delete?exNo=${exta.EX_NO }&userNo=${exta.POSTOWNER }"><button>게시글 삭제</button></a></td>
+		<td style="text-align: center;"><a href="/admin/user/history/extagram/delete?exNo=${exta.EX_NO }&userNo=${exta.POSTOWNER }" onclick="return confirm('게시글을 삭제하시겠습니까? 삭제된 게시글은 원상복구 할 수 없습니다.');"><button>게시글 삭제</button></a></td>
 	</tr>	
 </c:forEach>
 </tbody>
